Update README.md
Here's an explanation of the tool's purpose:
Encryption:

The encrpt_pass function takes a password as input from the user.
It encodes the password using the base64.b64encode() function, which converts the password into a Base64-encoded string.
The resulting encoded password is printed as output.
Decryption:

The decrpt_pass function expects an encoded password as input from the user.
It decodes the password using the base64.b64decode() function, which reverses the encoding process and retrieves the original password.
The decoded password is printed as output.
This tool can be useful in scenarios where you need to store or transmit passwords in a format that is easily readable and transferable. However, it's important to note that Base64 encoding is not a secure method for protecting sensitive information because it is easily reversible. Base64 encoding is mainly used for encoding data in situations where the data is not intended to be secret but requires a textual representation.

It's crucial to understand that this tool should not be considered a secure encryption or decryption mechanism for storing or transmitting passwords. For real-world security purposes, it is recommended to use more robust encryption algorithms and follow secure coding practices to protect sensitive information.